python env, DQN agent implemented / todo : Tensor.gather() has some problems

+++ src/main/java/tensor/Nd4jInit.java
@@ -5,6 +5,6 @@ import org.nd4j.linalg.factory.Nd4j;
 
 public class Nd4jInit {
     public static void configure() {
-        Nd4j.setDefaultDataTypes(DataType.FLOAT, DataType.FLOAT);
+        Nd4j.setDefaultDataTypes(DataType.DOUBLE, DataType.DOUBLE);
     }
 }
